config: extract compileOpts into compiler_opts.go and dedup lenient literals
Move the ~1,936-line `compileOpts` struct (with its doc comment) verbatim
out of the 2,761-line compiler.go into a dedicated sibling file
compiler_opts.go. The struct is a single, heavily-documented type
declaration — one lenient/strict severity flag per validation gate — that
dominated compiler.go; giving it its own file leaves compiler.go as the
lean compile-entry / orchestration surface (2,761 -> 566 lines) and
mirrors the #6441/#6442/#6443 per-aspect sibling-file splits.

Also dedup the two byte-identical ~125-field lenient composite literals
that CompileConfigLenient and CompileConfigForNodeLenient each inlined
into a single `lenientCompileOpts()` source-of-truth constructor. Both
call sites now pass `lenientCompileOpts()`; compileOpts is a value type,
so each call returns a fresh copy — behavior-identical to the former
inline literals, and a new tolerant-path downgrade is now added in one
place instead of two.

Pure code-motion / dedup, no logic change. The struct + doc comment moved
verbatim (byte-exact diff of the original compiler.go region vs the new
file: zero delta), and the constructor's field block is byte-identical to
both original twin literals (126 == 126 == 126 fields, zero delta).

Validation: go build ./..., go vet ./pkg/config/, gofmt clean; full
go test ./... GREEN (incl. the golden gate-ordering test and the #1373
pkg/dataplane retirement canary).
